Extract getConfirmationsForTx helpers to a separate module
